+++ authority-portal-frontend/src/app/common/layouts/portal-layout/sidebar/sidebar.component.ts
@@ -22,7 +22,7 @@ export class SidebarComponent {
           },
           {
             title: 'Profile',
-            icon: 'building-office',
+            icon: 'profile',
             rLink: '/my-organization/profile',
           },
           {