+++ authority-portal-frontend/src/app/pages/authority-organization-list-page/authority-organization-list-page.module.ts
@@ -5,6 +5,7 @@ import {NgxsModule} from '@ngxs/store';
 import {LoadingElementModule} from 'src/app/common/components/loading-element/loading-element.module';
 import {PipesAndDirectivesModule} from 'src/app/common/components/pipes-and-directives/pipes-and-directives.module';
 import {PortalLayoutModule} from 'src/app/common/layouts/portal-layout/portal-layout.module';
+import {SharedModule} from 'src/app/shared/shared.module';
 import {AuthorityOrganizationListPageComponent} from './authority-organization-list-page/authority-organization-list-page.component';
 import {AuthorityOrganizationListPageStateImpl} from './state/authority-organization-list-page-state-impl';
 
@@ -22,6 +23,7 @@ import {AuthorityOrganizationListPageStateImpl} from './state/authority-organiza
     PortalLayoutModule,
     PipesAndDirectivesModule,
     LoadingElementModule,
+    SharedModule,
   ],
 })
 export class AuthorityOrganizationListPageModule {}