+++ authority-portal-frontend/src/app/pages/authority-organization-list-page/state/authority-organization-list-page-actions.ts
@@ -3,3 +3,11 @@ const tag = 'AuthorityOrganizationListPage';
 export class RefreshOrganizations {
   static readonly type = `[${tag}] Refresh Organizations`;
 }
+
+export class ShowOrganizationDetail {
+  static readonly type = `[${tag}]  Show Organization Details Slider`;
+}
+
+export class CloseOrganizationDetail {
+  static readonly type = `[${tag}]  Close Organization Details Slider`;
+}